vision_transformer/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
Pillow==10.0.0
tqdm==4.65.0
PyYAML==6.0
timm==0.9.2
albumentations==1.3.1
wandb==0.16.0
torchmetrics==1.2.0
grad-cam==1.4.8
segmentation-models-pytorch==0.3.3
ultralytics==8.0.200
einops==0.7.0
tensorboard==2.15.1
onnx==1.15.0
onnxruntime==1.16.3
